backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
onnxruntime==1.18.0
opencv-python==4.10.0.84
numpy>=1.24.0
Pillow>=10.0.0
piexif>=1.1.3
pydantic>=2.0.0
